requirements.txt
Flask~=2.1.1
rosservice~=1.15.13
rospy~=1.15.13
wadl-planner
six
pandas~=1.3.4